Responsive and Nav Fixes
There's still some issues with nav. Namely need to come up with design choices in regards to the toggle when the user hasn't scrolled down any. I added in a plain purple background to the collapse, which works well on the homepage but sticks out on other pages due to the gradient. Beyond this, the CTA still needs to be looked at for responsive and the modal needs to be finished.